add and remove methods and totalpotentialprofit method added to shop class

+++ wizard_management/Accessories.java
@@ -3,10 +3,10 @@ import behaviours.*;
 
 public abstract class Accessories implements Sellable {
 
-  double sale_price;
-  double buy_price;
+  int sale_price;
+  int buy_price;
 
-  public Accessories(double sale_price, double buy_price) {
+  public Accessories(int sale_price, int buy_price) {
     this.sale_price = sale_price;
     this.buy_price = buy_price;
   }
